@@ -1,6 +1,6 @@
 {
   "name": "react-native-ui-lib",
-  "version": "3.3.132",
+  "version": "4.0.0",
   "main": "index.js",
   "typings": "typings/index.d.ts",
   "author": "Ethan Sharabi <ethan.shar@gmail.com>",
@@ -15,7 +15,7 @@
   },
   "scripts": {
     "start": "watchman watch-del-all && export DEV_MODE=true && react-native start",
-    "ios": "react-native run-ios --simulator='iPhone 7'",
+    "ios": "react-native run-ios",
     "android": "react-native run-android",
     "pretest": "npm run lint",
     "test": "jest",
@@ -38,28 +38,25 @@
     "build:typings": "rm -rf ./typings && ./scripts/build-typings.sh && tsc"
   },
   "dependencies": {
-    "@react-native-community/blur": "^3.3.1",
     "babel-plugin-transform-inline-environment-variables": "^0.0.2",
     "color": "^3.1.0",
     "commons-validator-js": "^1.0.237",
     "hoist-non-react-statics": "^3.0.0",
     "lodash": "^4.0.0",
+    "memoize-one": "^5.0.5",
     "prop-types": "^15.5.10",
     "react-native-animatable": "^1.1.0",
-    "react-native-gesture-handler": "1.3.0",
-    "react-native-interactable": "1.0.0",
-    "react-native-reanimated": "1.1.0",
-    "react-native-text-size": "^2.0.4",
-    "semver": "^5.5.0",
-    "url-parse": "^1.2.0",
     "react-native-color": "0.0.10",
-    "memoize-one": "^5.0.5"
+    "react-native-text-size": "4.0.0-rc.1",
+    "semver": "^5.5.0",
+    "url-parse": "^1.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.4.0",
     "@babel/preset-env": "^7.3.1",
     "@babel/runtime": "^7.4.2",
-    "@react-native-community/async-storage": "1.2.4",
+    "@react-native-community/async-storage": "^1.6.2",
+    "@react-native-community/blur": "^3.4.1",
     "@react-native-community/eslint-config": "latest",
     "@types/prop-types": "^15.5.3",
     "@types/react-native": "^0.55.21",
@@ -73,20 +70,28 @@
     "jest": "^24.5.0",
     "metro-react-native-babel-preset": "^0.53.1",
     "mocha": "^5.0.0",
-    "react": "16.8.3",
+    "react": "16.9.0",
+    "prettier-eslint": "^9.0.0",
     "react-autobind": "^1.0.6",
     "react-dom": "^15.4.2",
     "react-js-to-ts": "^1.4.0",
-    "react-native": "^0.59.3",
+    "react-native": "0.61.4",
     "react-native-keyboard-aware-scrollview": "^2.0.0",
-    "react-native-keyboard-tracking-view": "^5.3.2",
-    "react-native-navigation": "2.13.1",
+    "react-native-keyboard-tracking-view": "^5.6.1",
+    "react-native-navigation": "^3.4.0",
+    "react-native-gesture-handler": "1.5.2",
+    "react-native-reanimated": "1.4.0",
+    "react-native-interactable": "^2.0.0",
     "shell-utils": "^1.0.10",
     "typescript": "^2.9.2"
   },
   "peerDependencies": {
     "react": ">=16.0.0",
-    "react-native": ">=0.51.0"
+    "react-native": ">=0.51.0",
+    "react-native-gesture-handler": ">=1.5.0",
+    "react-native-reanimated": ">=1.4.0",
+    "@react-native-community/blur": ">=3.4.1",
+    "react-native-interactable": ">=2.0.0"
   },
   "jest": {
     "preset": "react-native",
