@@ -1,26 +1,20 @@
+import _ from 'lodash';
 import React, {Component} from 'react';
 import {StyleSheet} from 'react-native';
-import {Colors, View, Text} from 'react-native-ui-lib'; //eslint-disable-line
+import {View, Text, Card} from 'react-native-ui-lib'; //eslint-disable-line
 
 export default class PlaygroundScreen extends Component {
-  
-  constructor(props) {
-    super(props);
-    this.state = {};
-  }
-
-  componentDidMount() {}
-
   render() {
     return (
-      <View flex bg-dark80 center style={styles.container}>
-        <Text>Unicorn Playground Screen</Text>
+      <View center bg-dark80 flex>
+        <Card height={100} center padding-20>
+          <Text text50>Playground Screen</Text>
+        </Card>
       </View>
     );
   }
 }
 
 const styles = StyleSheet.create({
-  container: {
-  }
+  container: {}
 });
