@@ -1,8 +1,9 @@
 import _ from 'lodash';
 import React, {Component} from 'react';
 import {StyleSheet, ScrollView} from 'react-native';
-import {Constants, Colors, View, Card, Button, Text, Image} from 'react-native-ui-lib'; //eslint-disable-line
+import {Constants, Colors, View, Card, Button, Text, Image} from 'react-native-ui-lib'; // eslint-disable-line
 import posts from '../../data/posts';
+
 
 const featureIcon = require('../../assets/icons/star.png');
 const shareIcon = require('../../assets/icons/share.png');
@@ -12,18 +13,19 @@ const cardImage2 = require('../../assets/images/empty-state.jpg');
 export default class CardsScreen extends Component {
   state = {
     selected1: true,
-    selected2: true,
+    selected2: true
   };
 
   render() {
     const {selected1, selected2} = this.state;
+
     return (
       <View>
         <Image
           style={StyleSheet.absoluteFillObject}
           source={{
             uri:
-              'https://images.pexels.com/photos/1005644/pexels-photo-1005644.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=750&w=1260',
+              'https://images.pexels.com/photos/1005644/pexels-photo-1005644.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=750&w=1260'
           }}
         />
 
@@ -59,7 +61,7 @@ export default class CardsScreen extends Component {
                 activeOpacity={1}
                 marginR-20
               >
-                <Card.Image height={'100%'} imageSource={cardImage} />
+                <Card.Image height={'100%'} imageSource={cardImage}/>
               </Card>
               <Card
                 height={120}
@@ -74,7 +76,7 @@ export default class CardsScreen extends Component {
                   borderWidth: 3,
                 }}
               >
-                <Card.Image height={'100%'} imageSource={cardImage} />
+                <Card.Image height={'100%'} imageSource={cardImage}/>
               </Card>
             </View>
 
@@ -110,8 +112,8 @@ export default class CardsScreen extends Component {
               )}
             </View>
 
-            <Card row height={160} style={{marginBottom: 15}} onPress={() => {}} enableBlur>
-              <Card.Image width={115} imageSource={cardImage} />
+            <Card row height={160} style={{marginBottom: 15}} onPress={() => {}} enableBlur borderRadius={0}>
+              <Card.Image width={115} imageSource={cardImage}/>
               <View padding-20 flex>
                 <Text text70 dark10>
                   You’re Invited!
@@ -119,7 +121,6 @@ export default class CardsScreen extends Component {
                 <Text text80 dark10>
                   222 Join Old The Town Barbershop Official Store. Download the Wix app to...
                 </Text>
-
                 <Text text90 dark50>
                   wix.to/A465c
                 </Text>
@@ -131,16 +132,14 @@ export default class CardsScreen extends Component {
                 <Text text70 dark10>
                   You’re Invited!
                 </Text>
-
                 <Text text80 dark10>
                   Join Old The Town Barbershop Official Store. Download the Wix app to...
                 </Text>
-
                 <Text text90 dark50>
                   wix.to/A465c
                 </Text>
               </View>
-              <Card.Image width={115} imageSource={cardImage} />
+              <Card.Image width={115} imageSource={cardImage}/>
             </Card>
 
             <Card style={{marginBottom: 15}} onPress={() => {}}>
@@ -148,19 +147,19 @@ export default class CardsScreen extends Component {
                 <Text text70 dark10>
                   You’re Invited!
                 </Text>
-
                 <Text text90 dark50>
                   join now
                 </Text>
               </View>
-              <Card.Image height={120} imageSource={cardImage2} />
+              <Card.Image height={120} imageSource={cardImage2}/>
             </Card>
 
             {_.map(posts, (post, i) => {
               const statusColor = post.status === 'Published' ? Colors.green30 : Colors.orange30;
+              
               return (
                 <Card key={i} style={{marginBottom: 15}} onPress={() => console.log('press on a card')}>
-                  <Card.Image height={160} imageSource={post.coverImage} />
+                  <Card.Image height={160} imageSource={post.coverImage}/>
 
                   <View padding-20>
                     <Text text40 color={Colors.dark10}>
@@ -182,8 +181,8 @@ export default class CardsScreen extends Component {
                         {post.likes} Likes
                       </Text>
                       <View row right>
-                        <Button style={{marginRight: 10}} text90 link iconSource={featureIcon} label="Feature" />
-                        <Button text90 link iconSource={shareIcon} label="Share" />
+                        <Button style={{marginRight: 10}} text90 link iconSource={featureIcon} label="Feature"/>
+                        <Button text90 link iconSource={shareIcon} label="Share"/>
                       </View>
                     </View>
                   </View>
