@@ -1,65 +1,78 @@
 import React, {Component} from 'react';
 import {StyleSheet, ScrollView} from 'react-native';
-import {View, Text, Image} from 'react-native-ui-lib'; // eslint-disable-line
-
+import {View, Text, Image, Card, Constants} from 'react-native-ui-lib'; // eslint-disable-line
 
 const image = require('../../assets/images/card-example.jpg');
-// const uri = 'https://static.pexels.com/photos/50721/pencils-crayons-colourful-rainbow-50721.jpeg';
-// const icon = require('../../assets/icons/tags.png');
+const uri = {
+  uri: 'https://images.pexels.com/photos/140234/pexels-photo-140234.jpeg?auto=compress&cs=tinysrgb&dpr=3&h=750&w=1260'
+};
 
 export default class OverlaysScreen extends Component {
-  
+  getImageWithOverlay = type => {
+    return (
+      <View centerH>
+        <Text dark10>{type}</Text>
+        <Image style={styles.image} source={image} overlayType={type}/>
+      </View>
+    );
+  };
+
   render() {
     return (
       <ScrollView showsVerticalScrollIndicator={false}>
         <View flex padding-20>
-          <Text text40 marginB-40>Image Overlays</Text>
-          <View row>
-            <View centerH>
-            <Text dark10>VERTICAL</Text>
+          <Text text40 marginB-40>
+            Image Overlays
+          </Text>
+          <View row centerH>
+            {this.getImageWithOverlay(Image.overlayTypes.VERTICAL)}
+            {this.getImageWithOverlay(Image.overlayTypes.SOLID)}
+          </View>
+          <View row centerH>
+            {this.getImageWithOverlay(Image.overlayTypes.TOP)}
+            {this.getImageWithOverlay(Image.overlayTypes.BOTTOM)}
+          </View>
+
+          <Text text60BO marginB-10 marginT-20>
+            Image Overlays (absolute)
+          </Text>
+          <View style={styles.overlayImageAbsoluteContainer}>
             <Image
-              style={styles.image}
+              style={styles.overlayImageAbsoluteVertical}
               source={image}
               overlayType={Image.overlayTypes.VERTICAL}
             />
-            </View>
-            <View centerH>
-              <Text dark10>SOLID</Text>
-              <Image
-                style={styles.image}
-                source={image}
-                overlayType={Image.overlayTypes.SOLID}
-              />
+            <Image style={styles.overlayImageAbsoluteSolid} source={image} overlayType={Image.overlayTypes.SOLID}/>
+            <Image style={styles.overlayImageAbsoluteTop} source={image} overlayType={Image.overlayTypes.TOP}/>
+            <Image style={styles.overlayImageAbsoluteBottom} source={image} overlayType={Image.overlayTypes.BOTTOM}/>
           </View>
-          </View>
-          <View row>
-            <View centerH>
-              <Text dark10>TOP</Text>
-              <Image
-                style={styles.image}
-                source={image}
-                overlayType={Image.overlayTypes.TOP}
-              />
+
+          <Text text60BO marginB-10 marginT-20>
+            Card.Image Overlay
+          </Text>
+          <Card marginB-s4 padding-s2 borderRadius={4} row height={106}>
+            <Card.Image style={styles.cardImage} imageSource={image} overlayType={Image.overlayTypes.VERTICAL}/>
+            <View flex marginL-s2 centerV>
+              <Text text50>{Image.overlayTypes.VERTICAL}</Text>
+              <Text text70>Ipsum nostrud officia deserunt irure eu.</Text>
             </View>
-            <View centerH>
-              <Text dark10>BOTTOM</Text>
-              <Image
-                style={styles.image}
-                source={image}
-                overlayType={Image.overlayTypes.BOTTOM}
-              />
-            </View>
-          </View>
+          </Card>
         </View>
-        {/* <Image
-          source={icon}
-          overlayType={Image.overlayTypes.VERTICAL}
-          style={{margin: 10}}
-        /> */}
-        <Image
-          source={image}
-          overlayType={Image.overlayTypes.VERTICAL}
-        />
+
+        <Text text60BO margin-10>
+          Image Overlay (cover, remote)
+        </Text>
+        <Image cover source={uri} overlayType={Image.overlayTypes.VERTICAL}/>
+
+        <Text text60BO margin-10>
+        Image Overlay (cover, asset)
+        </Text>
+        <Image cover source={image} overlayType={Image.overlayTypes.VERTICAL}/>
+
+        <Text text60BO margin-10>
+        Image Overlay (fill, specific height)
+        </Text>
+        <Image source={image} overlayType={Image.overlayTypes.VERTICAL} style={styles.imageFillContainer}/>
       </ScrollView>
     );
   }
@@ -67,8 +80,52 @@ export default class OverlaysScreen extends Component {
 
 const styles = StyleSheet.create({
   image: {
-    width: 150, 
-    height: 150, 
+    width: 150,
+    height: 150,
     margin: 10
+  },
+  overlayImageAbsoluteContainer: {
+    width: Constants.screenWidth,
+    height: 400,
+    marginLeft: -20
+  },
+  overlayImageAbsoluteVertical: {
+    position: 'absolute',
+    top: 10,
+    left: 0,
+    width: 150,
+    height: 150,
+    margin: 5
+  },
+  overlayImageAbsoluteSolid: {
+    position: 'absolute',
+    top: 0,
+    right: 10,
+    width: 150,
+    height: 150,
+    margin: 5
+  },
+  overlayImageAbsoluteTop: {
+    position: 'absolute',
+    bottom: 0,
+    left: 10,
+    width: 150,
+    height: 150,
+    margin: 5
+  },
+  overlayImageAbsoluteBottom: {
+    position: 'absolute',
+    bottom: 10,
+    right: 0,
+    width: 150,
+    height: 150,
+    margin: 5
+  },
+  cardImage: {
+    width: 100,
+    borderRadius: 4
+  },
+  imageFillContainer: {
+    height: 100
   }
 });
