@@ -1,8 +1,9 @@
 package com.wix.reactnativeuilib.textinput;
 
-import android.support.annotation.Nullable;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.annotation.Nullable;
 
 import com.facebook.react.views.textinput.ReactEditText;
 
