@@ -14,8 +14,8 @@ import CardItem from './CardItem';
 import CardImage from './CardImage';
 import Assets from '../../assets';
 
-const DEFAULT_BORDER_RADIUS = BorderRadiuses.br40;
 
+const DEFAULT_BORDER_RADIUS = BorderRadiuses.br40;
 const DEFAULT_SELECTION_PROPS = {
   borderWidth: 2,
   color: Colors.blue30,
@@ -147,6 +147,7 @@ class Card extends PureBaseComponent {
 
   get elevationStyle() {
     const {elevation, enableShadow} = this.getThemeProps();
+    
     if (enableShadow) {
       return {elevation: elevation || 2};
     }
@@ -154,6 +155,7 @@ class Card extends PureBaseComponent {
 
   get shadowStyle() {
     const {enableShadow} = this.getThemeProps();
+    
     if (enableShadow) {
       return this.styles.containerShadow;
     }
@@ -161,6 +163,7 @@ class Card extends PureBaseComponent {
 
   get blurBgStyle() {
     const {enableBlur} = this.getThemeProps();
+    
     if (Constants.isIOS && enableBlur) {
       return {backgroundColor: Colors.rgba(Colors.white, 0.85)};
     } else {
@@ -168,10 +171,15 @@ class Card extends PureBaseComponent {
     }
   }
 
-  renderSelection() {
-    const {selectionOptions, borderRadius, selected} = this.getThemeProps();
-    const {animatedSelected} = this.state;
+  get borderRadius() {
+    const {borderRadius} = this.getThemeProps();
+    
+    return borderRadius === undefined ? DEFAULT_BORDER_RADIUS : borderRadius;
+  }
 
+  renderSelection() {
+    const {selectionOptions, selected} = this.getThemeProps();
+    const {animatedSelected} = this.state;
     const selectionColor = _.get(selectionOptions, 'color', DEFAULT_SELECTION_PROPS.color);
 
     if (_.isUndefined(selected)) {
@@ -183,7 +191,7 @@ class Card extends PureBaseComponent {
         style={[
           this.styles.selectedBorder,
           {borderColor: selectionColor},
-          borderRadius && {borderRadius},
+          {borderRadius: this.borderRadius},
           {opacity: animatedSelected}
         ]}
         pointerEvents="none"
@@ -196,13 +204,12 @@ class Card extends PureBaseComponent {
   }
 
   renderChildren() {
-    const {borderRadius} = this.getThemeProps();
     const children = React.Children.map(this.props.children, (child, index) => {
       if (_.get(child, 'type.displayName') === CardImage.displayName) {
         const position = this.calcImagePosition(index);
         return React.cloneElement(child, {
           position,
-          borderRadius: borderRadius || DEFAULT_BORDER_RADIUS
+          borderRadius: this.borderRadius
         });
       }
       return child;
@@ -211,10 +218,10 @@ class Card extends PureBaseComponent {
   }
 
   render() {
-    const {onPress, style, containerStyle, borderRadius, enableBlur, ...others} = this.getThemeProps();
+    const {onPress, style, selected, containerStyle, enableBlur, ...others} = this.getThemeProps();
     const blurOptions = this.getBlurOptions();
     const Container = onPress ? TouchableOpacity : View;
-    const brRadius = borderRadius || DEFAULT_BORDER_RADIUS;
+    const brRadius = this.borderRadius;
 
     return (
       <Container
@@ -230,6 +237,7 @@ class Card extends PureBaseComponent {
         onPress={onPress}
         delayPressIn={10}
         activeOpacity={0.6}
+        accessibilityState={{selected}}
         {...others}
         ref={this.setRef}
       >
@@ -244,14 +252,16 @@ class Card extends PureBaseComponent {
   }
 }
 
-function createStyles({width, height, borderRadius = DEFAULT_BORDER_RADIUS, selectionOptions}) {
+function createStyles({width, height, borderRadius, selectionOptions}) {
   const selectionOptionsWithDefaults = _.merge(DEFAULT_SELECTION_PROPS, selectionOptions);
+  const brRadius = borderRadius === undefined ? DEFAULT_BORDER_RADIUS : borderRadius;
+
   return StyleSheet.create({
     container: {
       width,
       height,
       overflow: 'visible',
-      borderRadius
+      borderRadius: brRadius
     },
     containerShadow: {
       // sh30 bottom
@@ -262,7 +272,7 @@ function createStyles({width, height, borderRadius = DEFAULT_BORDER_RADIUS, sele
     },
     blurView: {
       ...StyleSheet.absoluteFillObject,
-      borderRadius
+      borderRadius: brRadius
     },
     selectedBorder: {
       ...StyleSheet.absoluteFillObject,
